Poprawka zapisu ilosciCoinów

+++ Projekt Unity/TowerDefense/Assets/Scripts/NPC/SkryptyPomocnicze/PomocniczeFunkcje.cs
@@ -843,7 +843,7 @@ public static class PomocniczeFunkcje
 
             fs.Close();
 
-            ManagerGryScript.iloscCoinów = (ds.ilośćMonet <= 50) ? (ushort)50 : (ushort)ds.ilośćMonet;
+            ManagerGryScript.iloscCoinów = (ds.ilośćMonet <= 50) ? (int)50 : (int)ds.ilośćMonet;
             odblokowaneEpoki = ds._odblokowanieEpoki;
             odblokowanyPoziomEpoki = ds._odblokowanyPoziomEpoki;
             managerGryScript.hpIdx = (ushort)ds.poziomHP;
